+++ src/integrationTest/java/org/opensearch/security/plugin/SystemIndexPlugin1.java
@@ -29,6 +29,7 @@ import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.rest.RestController;
 import org.opensearch.rest.RestHandler;
 import org.opensearch.script.ScriptService;
+import org.opensearch.security.identity.TransportActionDependencies;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.watcher.ResourceWatcherService;
